Add support for SIP Voice Mail
The modern android way of handling voicemail, via the Dialer, is
closely tied to the SIM subscriber id.  Since Sip doesn't have this
(and has no use for the SIM subscription manager) we have to fall back
to the old way of singalling voicemail via a direct notification.
re-open this channel in the notification manager and use the incoming
call intent to signal the presence of voicemail.

Change-Id: I5e2cb6fdaee09740c02eab64d1b52279b72d0944
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>