make "Make calls with" option persist for SIP
Currently the "Make calls with" option is trashed on each phone reboot
because the subscription manager overwrites it each time.  However, if
the user has selected to make calls with a SIP account, we'd rather
not have the sim manager keep switching it back again.  The main
problem is that SIP accounts aren't started until first login because
their configuration is stored in encrypted files, so at that point,
the default subscription manager selection needs to be updated.

We do this by intercepting the user preference selection of calling
account via SipUtils and updating the sip profiles database to mark
the account as an outgoing phone account, while removing the flag from
all other sip profiles (ensuring there's only one such account
marked).  On SIP startup, if an account is marked like this, we set it
as the default outgoing phone account.

Change-Id: I0ae103296d3c0043dcc6fb6949d782708e8ddc51
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>